ax/llamaindex-py/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
llama-index-core>=0.12.0
llama-index-llms-anthropic>=0.6.0
chromadb>=0.6.0
pydantic>=2.0.0
arize-otel>=0.5.0
openinference-instrumentation-llama-index>=0.1.0
